requirements.txt
websocket-client
requests
python-dotenv
urllib3
tldextract
